put seat and input config in their own files
